requirements-dev.txt
pytest
requests
python-docx
SQLAlchemy
pydantic
pydantic-settings
reportlab
pypdf
